feat: 피드, 모임 안내 panel 분리
