+++ src/components/page/meetingDetail/FeedPanel.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const FeedPanel = () => {
+  return <div>피드</div>;
+};
+
+export default FeedPanel;
